Refactoring - SelectionMenu is the main component.
